Refactored transformer, compiler, and partials processor
