+++ lib/index.ts
@@ -1,21 +1,20 @@
 import type { Plugin } from 'rollup';
-import HandlebarsCompiler from './handlebars-compiler';
+import HandlebarsTransformer from './handlebars-transformer';
 
 import { HandlebarsPluginOptions } from './types/plugin-options';
-import path from 'path';
 
 export default function handlebarsCompilerPlugin(handlebarsPluginOptions?: HandlebarsPluginOptions): Plugin {
 
-	const compiler = new HandlebarsCompiler(handlebarsPluginOptions)
+	const hbsTransformer = new HandlebarsTransformer(handlebarsPluginOptions)
 
 	return {
 		name: 'handlebars-compiler',
 
 		transform(source, id) {
 			if (/\.(hbs|handlebars)/.test(id)) {
-				const output = compiler.tramsformHbs(source, id);
+				const output = hbsTransformer.transform(source, id);
 				const existingWatchFiles = this.getWatchFiles()
-				const watchFiles = compiler.getWatchFiles(existingWatchFiles);
+				const watchFiles = hbsTransformer.getWatchFiles(existingWatchFiles);
 				watchFiles.forEach((file: string) => {
 					this.addWatchFile(file)
 				})
